Project/requirements.txt
Flask
Flask-SQLAlchemy
requests
Pillow
numpy
scikit-learn
werkzeug
ultralytics     
torch            
sentence-transformers
rapidfuzz        